refactor: new group modal refactored
